Fisica 4.0
Changed organization:
-One screen for Flat Exercises
-One screen for Ramp Exercises
-Add X Force Button
-Add Y Force Button
-Add Diagonal Force Button
-Respective Delete Buttons

Files and Others
-TextViews below each Ramp Image
-Deleted files which were no longer needed
-.java is far from being functional in both cases
-.xml is soon being completed for Flat Exercise

09/15/2016
