+++ app/build.gradle
@@ -8,8 +8,8 @@ android {
         applicationId "com.example.administrador.fisica"
         minSdkVersion 15
         targetSdkVersion 24
-        versionCode 9
-        versionName "3.2.1"
+        versionCode 10
+        versionName "4.0"
     }
     buildTypes {
         release {
