--- app/src/main/java/com/example/administrador/fisica/FlatCombinedActivity.java
@@ -1,13 +1,0 @@
-package com.example.administrador.fisica;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class FlatCombinedActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_flat_combined);
-    }
-}
