+++ app/src/main/res/layout/activity_basic.xml
@@ -2,8 +2,6 @@
 <LinearLayout android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".FlatStraightActivity"
     android:orientation="vertical">
 
 <android.support.v7.widget.Toolbar
